+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0] - Unreleased
+### Added
+- Added the ability to do initial authentication with username/password
+  or with Kerberos tickets.
+### Deprecated
+- Passing the certfile/cert password as their own arguments is now deprecated.
+  Use the new `auth_type` argument to pass authentication information.
+### New contributors
+- Nickolai Zeldovich (zeldovich)
+
 ## [0.3.1] - 2022-02-23
 ### Added
 - Added the ability to load a wider variety of Bearer tokens (Tim tickets)